limit the led freq. and fix abs_pose bug

+++ src/buzz/buzzkh4_closures.h
@@ -25,7 +25,7 @@ extern int buzzkh4_set_led(buzzvm_t vm);
 extern int buzzkh4_set_led_freq(buzzvm_t vm);
 //extern void set_single_led(int r, int g, int b, int n);
 extern void turnon_led(uint8_t on);
-extern int get_led_freq();
+extern long get_led_freq();
 
 /*
  * Updates battery information in Buzz
